+++ src/gui/channel_messages.rs
@@ -262,6 +262,7 @@ pub mod imp {
             let obj = self.obj();
             let manager = obj.manager();
             let attachment = crate::backend::Attachment::from_texture(texture, &manager);
+
             self.append_attachment(attachment);
             obj.notify("has-attachments");
         }